backend/pyproject.toml
[project]
name = "cadlift-backend"
version = "0.1.0"
description = "CADLift FastAPI backend"
authors = [{name = "CADLift"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.24",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "passlib[argon2]>=1.7",
    "argon2-cffi>=23.1",
    "python-jose[cryptography]>=3.3",
    "email-validator>=2.1",
    "python-multipart>=0.0.9",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "ezdxf>=1.3",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "trimesh>=4.9",
    "transformers>=4.35.0",
    "httpx>=0.28",
    "typing-extensions>=4.8",
    "cadquery>=2.6",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "black>=24.3",
    "ruff>=0.3",
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uvicorn]
app = "app.main:app"
reload = true
host = "0.0.0.0"
port = 8000
